+++ src/Resources/Insights/PerformanceDetail.php
@@ -1,0 +1,45 @@
+<?php
+namespace Budgetlens\BolRetailerApi\Resources\Insights;
+
+use Budgetlens\BolRetailerApi\Resources\BaseResource;
+use Budgetlens\BolRetailerApi\Resources\Insights\Performance\Norm;
+use Budgetlens\BolRetailerApi\Resources\Insights\Performance\Score;
+
+class PerformanceDetail extends BaseResource
+{
+    public $week;
+    public $year;
+    public $score;
+    public $norm;
+
+    public function setPeriodAttribute($value): self
+    {
+        $this->week = (int)$value->week ?? null;
+        $this->year = (int)$value->year ?? null;
+
+        return $this;
+    }
+
+    public function setScoreAttribute($value): self
+    {
+        $this->score = new Score($value);
+
+        return $this;
+    }
+
+    public function setNormAttribute($value): self
+    {
+        $this->norm = new Norm($value);
+
+        return $this;
+    }
+
+    public function toArray(): array
+    {
+        return collect(parent::toArray())
+            ->reject(function ($value) {
+                return is_null($value);
+            })
+            ->all();
+    }
+}