+++ src/Resources/Insights/Search/RelatedTerm.php
@@ -1,0 +1,10 @@
+<?php
+namespace Budgetlens\BolRetailerApi\Resources\Insights\Search;
+
+use Budgetlens\BolRetailerApi\Resources\BaseResource;
+
+class RelatedTerm extends BaseResource
+{
+    public $total;
+    public $searchTerm;
+}